# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/brands/{brandId}/sales/sales-by-sku": {
      "get": {
        "operationId": "SalesController_getSalesBySkuData",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "search by sku",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SKUResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-sku-graph": {
      "get": {
        "operationId": "SalesController_getSalesBySkuGraph",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-product": {
      "get": {
        "operationId": "SalesController_getSalesByProduct",
        "parameters": [
          {
            "name": "brandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-week-summary": {
      "get": {
        "operationId": "SalesController_getSalesByWeekSummary",
        "parameters": [
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-week-data": {
      "get": {
        "operationId": "SalesController_getSalesByWeekData",
        "parameters": [
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-month-summary": {
      "get": {
        "operationId": "SalesController_getSalesByMonthSummary",
        "parameters": [
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-by-month-data": {
      "get": {
        "operationId": "SalesController_getSalesByMonthData",
        "parameters": [
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-detail": {
      "get": {
        "operationId": "SalesController_getSalesDetail",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "week",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-graph-data": {
      "get": {
        "operationId": "SalesController_getSalesGraphData",
        "parameters": [
          {
            "name": "graph_filter_type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sku",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "child_asins",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "parent_asins",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-asins": {
      "get": {
        "operationId": "SalesController_getAsins",
        "parameters": [
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "years",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-callout-data": {
      "get": {
        "operationId": "SalesController_getCalloutData",
        "parameters": [
          {
            "name": "weeks",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-callout-data-monthly": {
      "get": {
        "operationId": "SalesController_getCalloutDataMonthly",
        "parameters": [
          {
            "name": "months",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-selected-callout": {
      "get": {
        "operationId": "SalesController_getSelectedCalloutData",
        "parameters": [
          {
            "name": "week",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/brands/{brandId}/sales/sales-selected-callout-monthly": {
      "get": {
        "operationId": "SalesController_getSelectedCalloutDatMonthly",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Vendo API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductDetailsDto": {
        "type": "object",
        "properties": {
          "child_asin": {
            "type": "string"
          },
          "parent_asin": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "item_names": {
            "type": "string"
          },
          "item_descriptions": {
            "type": "string"
          },
          "seller_skus": {
            "type": "string"
          },
          "image_urls": {
            "type": "string"
          },
          "product_ids": {
            "type": "string"
          },
          "astr_units_ordered_sum": {
            "type": "string"
          },
          "ordered_product_sales_sum": {
            "type": "string"
          },
          "astr_buy_box_percentage_sum": {
            "type": "string"
          },
          "unit_session_percentage_sum": {
            "type": "string"
          },
          "astr_sessions_sum": {
            "type": "string"
          },
          "astr_page_views_sum": {
            "type": "string"
          },
          "astr_session_percentage_sum": {
            "type": "string"
          },
          "total_order_items_sum": {
            "type": "string"
          },
          "astr_page_view_percentage_sum": {
            "type": "string"
          },
          "astr_buy_box_percentage_avg": {
            "type": "string"
          },
          "unit_session_percentage_avg": {
            "type": "string"
          },
          "astr_session_percentage_avg": {
            "type": "string"
          },
          "astr_page_view_percentage_avg": {
            "type": "string"
          },
          "astr_child_asin_count": {
            "type": "number"
          }
        },
        "required": [
          "child_asin",
          "parent_asin",
          "sku",
          "title",
          "item_names",
          "item_descriptions",
          "seller_skus",
          "image_urls",
          "product_ids",
          "astr_units_ordered_sum",
          "ordered_product_sales_sum",
          "astr_buy_box_percentage_sum",
          "unit_session_percentage_sum",
          "astr_sessions_sum",
          "astr_page_views_sum",
          "astr_session_percentage_sum",
          "total_order_items_sum",
          "astr_page_view_percentage_sum",
          "astr_buy_box_percentage_avg",
          "unit_session_percentage_avg",
          "astr_session_percentage_avg",
          "astr_page_view_percentage_avg",
          "astr_child_asin_count"
        ]
      },
      "ProductSummaryDto": {
        "type": "object",
        "properties": {
          "totalSkuCount": {
            "type": "string"
          },
          "totalUnitOrdered": {
            "type": "string"
          },
          "totalOrderedProductSales": {
            "type": "string"
          },
          "avgBuyBox": {
            "type": "string"
          },
          "avgUnitSession": {
            "type": "string"
          },
          "totalSession": {
            "type": "string"
          },
          "totalPageViews": {
            "type": "string"
          },
          "totalSessionPercentage": {
            "type": "string"
          },
          "totalOrderItems": {
            "type": "string"
          },
          "avgPageViewPercentage": {
            "type": "string"
          }
        },
        "required": [
          "totalSkuCount",
          "totalUnitOrdered",
          "totalOrderedProductSales",
          "avgBuyBox",
          "avgUnitSession",
          "totalSession",
          "totalPageViews",
          "totalSessionPercentage",
          "totalOrderItems",
          "avgPageViewPercentage"
        ]
      },
      "SKUResponseDto": {
        "type": "object",
        "properties": {
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDetailsDto"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/ProductSummaryDto"
          }
        },
        "required": [
          "details",
          "summary"
        ]
      },
...
    }
  }
}